# contracts/openapi.yaml
...
openapi: 3.1.0
openapi: 3.1.0
info:
  title: Retail Store API
  description: |-
    API REST del e-commerce Retail Store.

    Los errores siguen RFC 9457 (application/problem+json) con las extensiones code y correlationId. El cliente se ramifica por code, nunca por detail.
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/cuenta/favoritos/{productoId}:
    put:
      tags:
      - Favoritos (tienda)
      summary: Marcar un producto como favorito. Idempotente
      operationId: agregar
      parameters:
      - name: productoId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      security:
      - tokenTienda: []
    delete:
      tags:
      - Favoritos (tienda)
      summary: Quitarlo de favoritos. Idempotente
      operationId: quitar
      parameters:
      - name: productoId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      security:
      - tokenTienda: []
  /api/v1/cuenta/direcciones/{id}:
    put:
      tags:
      - Direcciones (tienda)
      summary: Editar una dirección
      operationId: actualizar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DireccionPeticion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DireccionRespuesta"
      security:
      - tokenTienda: []
    delete:
      tags:
      - Direcciones (tienda)
      summary: "Eliminar una dirección. Es lógica (RN-086): las órdenes antiguas siguen\
        \ explicándose"
      operationId: eliminar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      security:
      - tokenTienda: []
  /api/v1/admin/subcategorias/{id}:
    get:
      tags:
      - Subcategorías (panel)
      summary: Detalle de una subcategoría
      operationId: porId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubcategoriaRespuesta"
      security:
      - tokenPanel: []
    put:
      tags:
      - Subcategorías (panel)
      summary: Actualizar una subcategoría; admite moverla de categoría
      operationId: actualizar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarSubcategoriaPeticion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SubcategoriaRespuesta"
      security:
      - tokenPanel: []
    delete:
      tags:
      - Subcategorías (panel)
      summary: "Eliminación lógica: la fila se conserva con quién y cuándo"
      operationId: eliminar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      security:
      - tokenPanel: []
  /api/v1/admin/productos/{id}:
    get:
      tags:
      - Productos (panel)
      summary: Detalle de un producto por id
      operationId: porId_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoAdminRespuesta"
      security:
      - tokenPanel: []
    put:
      tags:
      - Productos (panel)
      summary: Actualizar un producto. El SKU no se puede cambiar
      operationId: actualizar_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarProductoPeticion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoAdminRespuesta"
      security:
      - tokenPanel: []
    delete:
      tags:
      - Productos (panel)
      summary: "Eliminación lógica: sale del catálogo y del panel, pero la fila se\
        \ conserva"
      operationId: eliminar_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      security:
      - tokenPanel: []
  /api/v1/admin/marcas/{id}:
    get:
      tags:
      - Marcas (panel)
      summary: Detalle de una marca
      operationId: porId_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MarcaRespuesta"
      security:
      - tokenPanel: []
    put:
      tags:
      - Marcas (panel)
      summary: Actualizar una marca
      operationId: actualizar_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarMarcaPeticion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MarcaRespuesta"
      security:
      - tokenPanel: []
    delete:
      tags:
      - Marcas (panel)
      summary: "Eliminación lógica: la fila se conserva con quién y cuándo"
      operationId: eliminar_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      security:
      - tokenPanel: []
  /api/v1/admin/cupones/{id}:
    get:
      tags:
      - Cupones (panel)
      summary: Detalle de un cupón
      operationId: porId_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CuponRespuesta"
      security:
      - tokenPanel: []
    put:
      tags:
      - Cupones (panel)
      summary: Actualizar un cupón. El código no cambia
      operationId: actualizar_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CuponPeticion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CuponRespuesta"
      security:
      - tokenPanel: []
    delete:
      tags:
      - Cupones (panel)
      summary: "Eliminación lógica. El código no se libera: aparece en órdenes pasadas"
      operationId: eliminar_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      security:
      - tokenPanel: []
  /api/v1/admin/categorias/{id}:
    get:
      tags:
      - Categorías (panel)
      summary: "Detalle de una categoría, con sus subcategorías"
      operationId: porId_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoriaRespuesta"
      security:
      - tokenPanel: []
    put:
      tags:
      - Categorías (panel)
      summary: Actualizar una categoría
      operationId: actualizar_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarCategoriaPeticion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoriaRespuesta"
      security:
      - tokenPanel: []
    delete:
      tags:
      - Categorías (panel)
      summary: "Eliminación lógica: la fila se conserva con quién y cuándo"
      operationId: eliminar_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
      security:
      - tokenPanel: []
  /api/v1/admin/administradores/{id}:
    put:
      tags:
      - Administradores
      summary: "Actualizar nombre, rol y estado"
      operationId: actualizar_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarAdministradorPeticion"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdministradorRespuesta"
      security:
      - tokenPanel: []
...
  /api/v1/cuenta/direcciones:
    get:
      tags:
      - Direcciones (tienda)
      summary: "Las direcciones del cliente, la predeterminada primero"
      operationId: listar
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DireccionRespuesta"
      security:
      - tokenTienda: []
    post:
      tags:
      - Direcciones (tienda)
      summary: Añadir una dirección. La primera queda predeterminada siempre
      operationId: crear
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DireccionPeticion"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DireccionRespuesta"
      security:
      - tokenTienda: []
  /api/v1/cuenta/direcciones/{id}/predeterminada:
    post:
      tags:
      - Direcciones (tienda)
      summary: Marcarla como predeterminada. Desmarca la anterior
      operationId: predeterminada
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DireccionRespuesta"
      security:
      - tokenTienda: []
...
  /api/v1/cuenta/favoritos:
    get:
      tags:
      - Favoritos (tienda)
      summary: "Los favoritos del cliente, del más reciente al más antiguo"
      operationId: listar_9
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductoResumenRespuesta"
      security:
      - tokenTienda: []
...
components:
  schemas:
...
    DireccionPeticion:
      type: object
      properties:
        distritoId:
          type: integer
          format: int32
        etiqueta:
          type: string
          maxLength: 40
          minLength: 0
        destinatario:
          type: string
          maxLength: 120
          minLength: 0
        telefono:
          type: string
          maxLength: 20
          minLength: 0
        calle:
          type: string
          maxLength: 200
          minLength: 0
        numero:
          type: string
          maxLength: 20
          minLength: 0
        referencia:
          type: string
          maxLength: 300
          minLength: 0
        codigoPostal:
          type: string
          maxLength: 12
          minLength: 0
        latitud:
          type: number
          maximum: 90
          minimum: -90
        longitud:
          type: number
          maximum: 180
          minimum: -180
        predeterminada:
          type: boolean
      required:
      - calle
      - destinatario
      - distritoId
      - etiqueta
      - telefono
    DireccionRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        etiqueta:
          type: string
        destinatario:
          type: string
        telefono:
          type: string
        calle:
          type: string
        numero:
          type: string
        referencia:
          type: string
        codigoPostal:
          type: string
        latitud:
          type: number
        longitud:
          type: number
        predeterminada:
          type: boolean
        distrito:
          $ref: "#/components/schemas/UbigeoRespuesta"
        provincia:
          $ref: "#/components/schemas/UbigeoRespuesta"
        departamento:
          $ref: "#/components/schemas/UbigeoRespuesta"
    UbigeoRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
    ActualizarSubcategoriaPeticion:
      type: object
      properties:
        categoriaId:
          type: integer
          format: int64
        nombre:
          type: string
          maxLength: 80
          minLength: 0
        descripcion:
          type: string
          maxLength: 500
          minLength: 0
        imagenId:
          type: integer
          format: int64
        orden:
          type: integer
          format: int32
        activa:
          type: boolean
      required:
      - categoriaId
      - nombre
    ArchivoRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
        urls:
          type: object
          additionalProperties:
            type: string
        textoAlt:
          type: string
        ancho:
          type: integer
          format: int32
        alto:
          type: integer
          format: int32
        bytes:
          type: integer
          format: int64
        nombreOriginal:
          type: string
        creadoEn:
          type: string
          format: date-time
    ReferenciaRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        slug:
          type: string
    SubcategoriaRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        slug:
          type: string
        descripcion:
          type: string
        imagen:
          $ref: "#/components/schemas/ArchivoRespuesta"
        orden:
          type: integer
          format: int32
        activa:
          type: boolean
        categoria:
          $ref: "#/components/schemas/ReferenciaRespuesta"
    ActualizarProductoPeticion:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 160
          minLength: 0
        descripcionCorta:
          type: string
          maxLength: 300
          minLength: 0
        descripcion:
          type: string
          maxLength: 5000
          minLength: 0
        subcategoriaId:
          type: integer
          format: int64
        marcaId:
          type: integer
          format: int64
        precio:
          type: number
        precioAnterior:
          type: number
        stock:
          type: integer
          format: int32
        destacado:
          type: boolean
        imagenIds:
          type: array
          items:
            type: integer
            format: int64
      required:
      - nombre
      - precio
      - subcategoriaId
    ImagenRespuesta:
      type: object
      properties:
        archivoId:
          type: integer
          format: int64
        miniatura:
          type: string
        tarjeta:
          type: string
        detalle:
          type: string
        original:
          type: string
        textoAlt:
          type: string
        ancho:
          type: integer
          format: int32
        alto:
          type: integer
          format: int32
    ProductoAdminRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sku:
          type: string
        nombre:
          type: string
        slug:
          type: string
        descripcionCorta:
          type: string
        descripcion:
          type: string
        precio:
          type: number
        precioAnterior:
          type: number
        porcentajeDescuento:
          type: integer
          format: int32
        stock:
          type: integer
          format: int32
        destacado:
          type: boolean
        activo:
          type: boolean
        calificacionPromedio:
          type: number
        calificacionConteo:
          type: integer
          format: int32
        vistas:
          type: integer
          format: int64
        imagenes:
          type: array
          items:
            $ref: "#/components/schemas/ImagenRespuesta"
        marca:
          $ref: "#/components/schemas/ReferenciaRespuesta"
        subcategoria:
          $ref: "#/components/schemas/ReferenciaRespuesta"
        categoria:
          $ref: "#/components/schemas/ReferenciaRespuesta"
        creadoEn:
          type: string
          format: date-time
        creadoPor:
          type: string
        actualizadoEn:
          type: string
          format: date-time
        actualizadoPor:
          type: string
        version:
          type: integer
          format: int32
    ActualizarMarcaPeticion:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 80
          minLength: 0
        descripcion:
          type: string
          maxLength: 500
          minLength: 0
        logoId:
          type: integer
          format: int64
        activa:
          type: boolean
      required:
      - nombre
    MarcaRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        slug:
          type: string
        descripcion:
          type: string
        logo:
          $ref: "#/components/schemas/ArchivoRespuesta"
        activa:
          type: boolean
    CuponPeticion:
      type: object
      properties:
        codigo:
          type: string
          maxLength: 40
          minLength: 0
          pattern: "[A-Za-z0-9._-]+"
        tipo:
          type: string
          minLength: 1
          pattern: PORCENTAJE|MONTO_FIJO
        valor:
          type: number
        subtotalMinimo:
          type: number
        iniciaEn:
          type: string
          format: date-time
        terminaEn:
          type: string
          format: date-time
        usosMaximos:
          type: integer
          format: int32
        activo:
          type: boolean
      required:
      - codigo
      - iniciaEn
      - terminaEn
      - tipo
      - valor
    CuponRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        codigo:
          type: string
        tipo:
          type: string
        valor:
          type: number
        subtotalMinimo:
          type: number
        iniciaEn:
          type: string
          format: date-time
        terminaEn:
          type: string
          format: date-time
        usosMaximos:
          type: integer
          format: int32
        usosActuales:
          type: integer
          format: int32
        activo:
          type: boolean
        vigente:
          type: boolean
    ActualizarCategoriaPeticion:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 80
          minLength: 0
        descripcion:
          type: string
          maxLength: 500
          minLength: 0
        imagenId:
          type: integer
          format: int64
        orden:
          type: integer
          format: int32
        activa:
          type: boolean
      required:
      - nombre
    CategoriaRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        slug:
          type: string
        descripcion:
          type: string
        imagen:
          $ref: "#/components/schemas/ArchivoRespuesta"
        orden:
          type: integer
          format: int32
        activa:
          type: boolean
        subcategorias:
          type: array
          items:
            $ref: "#/components/schemas/SubcategoriaRespuesta"
    ActualizarAdministradorPeticion:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 120
          minLength: 0
        rol:
          type: string
          minLength: 1
          pattern: ADMINISTRADOR|SUPERADMINISTRADOR
        activo:
          type: boolean
      required:
      - nombre
      - rol
    AdministradorRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        usuario:
          type: string
        nombre:
          type: string
        rol:
          type: string
        activo:
          type: boolean
        ultimoAccesoEn:
          type: string
          format: date-time
        creadoEn:
          type: string
          format: date-time
...
    ProductoResumenRespuesta:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        slug:
          type: string
        descripcionCorta:
          type: string
        precio:
          type: number
        precioAnterior:
          type: number
        porcentajeDescuento:
          type: integer
          format: int32
        hayStock:
          type: boolean
        destacado:
          type: boolean
        calificacionPromedio:
          type: number
        calificacionConteo:
          type: integer
          format: int32
        imagen:
          type: string
        textoAltImagen:
          type: string
        marca:
          $ref: "#/components/schemas/ReferenciaRespuesta"
        subcategoria:
          $ref: "#/components/schemas/ReferenciaRespuesta"
...
  securitySchemes:
    tokenPanel:
      type: http
      description: Token de acceso del panel. Se obtiene en POST /api/v1/admin/acceso.
      scheme: bearer
      bearerFormat: JWT
    tokenTienda:
      type: http
      description: Token de acceso del cliente. Se obtiene en POST /api/v1/cuenta/acceso
        o /api/v1/cuenta/google.
      scheme: bearer
      bearerFormat: JWT